Refactor MoE blocks
