+++ MaxText/configs/models/mixtral-8x7b.yml
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 # model config for mixtral-8x7b
+# tokenizer_path is assets/tokenizer.mistral-v1
 
 base_emb_dim: 4096
 base_num_query_heads: 32
@@ -28,4 +29,4 @@ normalization_layer_epsilon: 1.0e-5
 num_experts: 8
 num_experts_per_tok: 2
 rope_max_timescale: 1_000_000
-decoder_block: "mistral"
+decoder_block: "mixtral"
